LPS-77163 Consolidates id, elementClasses and spritemap logic and id and elementClasses tag attributes
