+++ modules/apps/foundation/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/servlet/taglib/soy/CheckboxTag.java
@@ -33,16 +33,8 @@ public class CheckboxTag extends BaseClayTag {
 		putValue("disabled", disabled);
 	}
 
-	public void setElementClasses(String elementClasses) {
-		putValue("elementClasses", elementClasses);
-	}
-
 	public void setHideLabel(Boolean hideLabel) {
 		putValue("hideLabel", hideLabel);
-	}
-
-	public void setId(String id) {
-		putValue("id", id);
 	}
 
 	public void setIndeterminate(Boolean indeterminate) {
