+++ modules/apps/foundation/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/servlet/taglib/soy/LabelTag.java
@@ -15,11 +15,6 @@
 package com.liferay.frontend.taglib.clay.servlet.taglib.soy;
 
 import com.liferay.frontend.taglib.clay.servlet.taglib.soy.base.BaseClayTag;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.util.Map;
 
 /**
  * @author Chema Balsas
@@ -30,32 +25,12 @@ public class LabelTag extends BaseClayTag {
 		super("label", "ClayLabel");
 	}
 
-	@Override
-	public int doStartTag() {
-		Map<String, Object> context = getContext();
-
-		if (Validator.isNull(context.get("spritemap"))) {
-			ThemeDisplay themeDisplay = (ThemeDisplay)request.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-			putValue(
-				"spritemap",
-				themeDisplay.getPathThemeImages().concat("/clay/icons.svg"));
-		}
-
-		return super.doStartTag();
-	}
-
 	public void setCloseable(Boolean closeable) {
 		putValue("closeable", closeable);
 	}
 
 	public void setHref(String href) {
 		putValue("href", href);
-	}
-
-	public void setId(String id) {
-		putValue("id", id);
 	}
 
 	public void setLabel(String label) {
@@ -68,10 +43,6 @@ public class LabelTag extends BaseClayTag {
 
 	public void setSize(String size) {
 		putValue("size", size);
-	}
-
-	public void setSpritemap(String spritemap) {
-		putValue("spritemap", spritemap);
 	}
 
 	public void setStyle(String style) {
